+++ @types/Sidebar.ts
@@ -1,5 +1,5 @@
-import {ElementType, FunctionComponent, ReactElement, SyntheticEvent} from 'react';
-import {Callback, Color, DefaultProps, IconProp} from './_shared';
+import { ElementType, FunctionComponent, ReactElement, SyntheticEvent } from 'react';
+import { Callback, Color, DefaultProps, IconProp } from './_shared';
 
 export type SidebarProps = DefaultProps & {
   isOpen: boolean,
@@ -10,10 +10,12 @@ export type SidebarProps = DefaultProps & {
   headerLogo?: ReactElement | ElementType,
   toggleIcon?: IconProp,
   showToggle?: boolean,
-  HeaderRender?: ElementType,
   orientation?: 'left' | 'right',
   type?: 'shrinkable' | 'offcanvas',
-  transitionType?: 'translate' | 'margin'
+  transitionType?: 'translate' | 'margin',
+  HeaderRender?: ElementType,
+  ElementRender?: ElementType,
+  NavRender?: ElementType,
 };
 
 export type SidebarItemProps = DefaultProps & {
@@ -23,7 +25,10 @@ export type SidebarItemProps = DefaultProps & {
   current?: boolean,
   onClick?: Callback<SyntheticEvent<MouseEvent>>,
   LinkRender?: ElementType,
+  ElementRender?: ElementType,
 };
+
+export type SidebarDividerProps = DefaultProps & { ElementRender?: ElementType };
 
 export type SidebarCollapsibleProps = DefaultProps & {
   text: string,
@@ -31,7 +36,17 @@ export type SidebarCollapsibleProps = DefaultProps & {
   current?: boolean,
   showOpen?: boolean,
   LinkRender?: ElementType,
+  ElementRender?: ElementType,
+  ListRender?: ElementType,
 };
+
+export type SidebarGroupProps = DefaultProps & {
+  text: string,
+  Icon?: IconProp,
+  ItemRender?: ElementType,
+  ElementRender?: ElementType,
+  ListRender?: ElementType,
+}
 
 /**
  * Declares the Sidebar functional component
@@ -39,7 +54,8 @@ export type SidebarCollapsibleProps = DefaultProps & {
 declare const Sidebar: FunctionComponent<SidebarProps>
   & { Item: FunctionComponent<SidebarItemProps> }
   & { Collapsible: FunctionComponent<SidebarCollapsibleProps> }
-  & { Divider: FunctionComponent<DefaultProps> };
+  & { Divider: FunctionComponent<SidebarDividerProps> }
+  & { Group: FunctionComponent<SidebarGroupProps> };
 
 
 export default Sidebar;